+++ mitra_models/src/invoices/helpers.rs
@@ -12,7 +12,7 @@ use super::queries::{
 };
 use super::types::{DbInvoice, InvoiceStatus};
 
-pub async fn invoice_forwarded(
+pub async fn local_invoice_forwarded(
     db_client: &mut impl DatabaseClient,
     invoice_id: &Uuid,
     payout_tx_id: &str,
@@ -32,7 +32,7 @@ pub async fn invoice_forwarded(
     Ok(invoice)
 }
 
-pub async fn invoice_reopened(
+pub async fn local_invoice_reopened(
     db_client: &mut impl DatabaseClient,
     invoice_id: &Uuid,
 ) -> Result<DbInvoice, DatabaseError> {
@@ -78,7 +78,7 @@ mod tests {
     use serial_test::serial;
     use mitra_utils::caip2::ChainId;
     use crate::database::test_utils::create_test_database;
-    use crate::invoices::queries::{create_invoice, create_remote_invoice};
+    use crate::invoices::queries::{create_local_invoice, create_remote_invoice};
     use crate::profiles::{
         queries::create_profile,
         types::ProfileCreateData,
@@ -91,7 +91,7 @@ mod tests {
 
     #[tokio::test]
     #[serial]
-    async fn test_invoice_forwarded_and_reopened() {
+    async fn test_local_invoice_forwarded_and_reopened() {
         let db_client = &mut create_test_database().await;
         let sender_data = ProfileCreateData {
             username: "sender".to_string(),
@@ -104,7 +104,7 @@ mod tests {
             ..Default::default()
         };
         let recipient = create_user(db_client, recipient_data).await.unwrap();
-        let invoice = create_invoice(
+        let invoice = create_local_invoice(
             db_client,
             &sender.id,
             &recipient.id,
@@ -119,7 +119,7 @@ mod tests {
         ).await.unwrap();
 
         let payout_tx_id = "12abcd";
-        let invoice = invoice_forwarded(
+        let invoice = local_invoice_forwarded(
             db_client,
             &invoice.id,
             payout_tx_id,
@@ -133,7 +133,7 @@ mod tests {
             InvoiceStatus::Completed,
         ).await.unwrap();
 
-        let invoice = invoice_reopened(
+        let invoice = local_invoice_reopened(
             db_client,
             &invoice.id,
         ).await.unwrap();